Менеджер браузеров, фиксы
