+++ src/main/java/pageobject/LoginPage.java
@@ -1,6 +1,6 @@
 package pageobject;
 
-import data.Account;
+import tools.data.Account;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.WebElement;
 import org.openqa.selenium.support.FindBy;
