+++ src/main/java/pageobject/RegisterPage.java
@@ -1,6 +1,6 @@
 package pageobject;
 
-import data.Account;
+import tools.data.Account;
 import org.openqa.selenium.Keys;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.WebElement;
